% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posit_versions.R
\name{audit_posit_version}
\alias{audit_posit_version}
\title{Audit Posit Server}
\usage{
audit_posit_version(server_version, type = c("connect", "workbench"))
}
\arguments{
\item{server_version}{Current Posit version}

\item{type}{Posit product of interest}
}
\description{
Used for side effect, i.e. printing to console
}
\examples{
audit_posit_version("2023.03.0", type = "connect")
}
