% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_version.R
\name{audit_server_version}
\alias{audit_server_version}
\title{Audit Posit Server}
\usage{
audit_server_version(server_version, type = c("connect", "workbench"))
}
\arguments{
\item{server_version}{Current Posit version}

\item{type}{Posit product of interest}
}
\description{
Used for side effect, i.e. printing to console
}
\examples{
audit_server_version("2023.03.0", type = "connect")
}
