% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/software-versions.R
\name{augment_installed}
\alias{augment_installed}
\title{Augments installed software columns}
\usage{
augment_installed(installed, verbose = TRUE)
}
\arguments{
\item{installed}{A tibble with columns software and installed_version}

\item{verbose}{Default TRUE. Out a {cli} summary}
}
\description{
Augments installed software columns
}
