% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_server_headers.R
\name{check_server_headers}
\alias{check_server_headers}
\title{Check server headers}
\usage{
check_server_headers(server)
}
\arguments{
\item{server}{URL of server}
}
\description{
In addition to the checks made by serverHeaders we also check
that Posit isn't in one of the headers.
}
