% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sys_deps.R
\name{check_sys_deps}
\alias{check_sys_deps}
\title{Checking System Dependencies}
\usage{
check_sys_deps(os_release, installed_libs, debug_level = 0:2)
}
\arguments{
\item{os_release}{Output from /etc/os-release}

\item{installed_libs}{Output get_installed_libs()}

\item{debug_level}{Integer - 0, 1, 2}
}
\description{
Many R packages require additional system dependencies.
This check, obtains a list of system libraries installed on the connect server, then
determines which R packages can't be installed.
}
\details{
Suppress parameter will be changed in a near-future MR to verbose levels
}
