% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_timezone.R
\name{check_timezone}
\alias{check_timezone}
\title{Check Timezone Correctly Set}
\usage{
check_timezone()
}
\value{
Invisibly, `TRUE` if the timezone is set, `FALSE` otherwise.
}
\description{
Checks that the system timezone can be determined via `Sys.timezone()`.
}
