% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto-helpers.R
\name{get_quarto_server_header}
\alias{get_quarto_server_header}
\alias{get_quarto_sys_deps}
\alias{get_quarto_software_versions}
\alias{get_quarto_deploy}
\alias{get_quarto_server_version_msg}
\title{Quarto Helper Functions}
\usage{
get_quarto_server_header(out)

get_quarto_sys_deps(out)

get_quarto_software_versions(out)

get_quarto_deploy(out)

get_quarto_server_version_msg(out, versions)
}
\arguments{
\item{out}{Output from check}

\item{versions}{A tibble with server versions}
}
\description{
In general not used by users.
}
\details{
The server versions are available as get_connect_versions()
or get_workbench_versions()
}
