% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suppress.R
\name{get_suppress}
\alias{get_suppress}
\alias{get_debug_level}
\title{Debugging helper functions}
\usage{
get_suppress(debug_level)

get_debug_level(debug_level)
}
\arguments{
\item{debug_level}{An integer, 0, 1, 2.}
}
\description{
Debugging helper functions
}
