% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{init_r6_checks}
\alias{init_r6_checks}
\alias{init_r6_check}
\title{R6 Helper functions}
\usage{
init_r6_checks(dir, file, pkg_name)

init_r6_check(export, dir, file, pkg_name)
}
\arguments{
\item{dir}{Directory of the config file}

\item{file}{Config file name}

\item{pkg_name}{E.g. uat.base description}

\item{export}{Name of the R6 class to initialise}
}
\description{
These functions are used by the health check function
to initialise and check
}
