% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6-logger.R
\name{logger}
\alias{logger}
\title{R6 logger class}
\description{
Not called directly, but is inherited by base_check.
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-logger-start_logger}{\code{logger$start_logger()}}
    \item \href{#method-logger-stop_logger}{\code{logger$stop_logger()}}
    \item \href{#method-logger-get_log}{\code{logger$get_log()}}
    \item \href{#method-logger-clone}{\code{logger$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-logger-start_logger"></a>}}
\if{latex}{\out{\hypertarget{method-logger-start_logger}{}}}
\subsection{\code{logger$start_logger()}}{
  Starts the logger
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{logger$start_logger()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-logger-stop_logger"></a>}}
\if{latex}{\out{\hypertarget{method-logger-stop_logger}{}}}
\subsection{\code{logger$stop_logger()}}{
  Stores the results of the check
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{logger$stop_logger(passed)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{passed}}{Logical, if skipped, the NA}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-logger-get_log"></a>}}
\if{latex}{\out{\hypertarget{method-logger-get_log}{}}}
\subsection{\code{logger$get_log()}}{
  Retrieve the log
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{logger$get_log()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Check log
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-logger-clone"></a>}}
\if{latex}{\out{\hypertarget{method-logger-clone}{}}}
\subsection{\code{logger$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{logger$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
