% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_all_versions.R
\name{update_all_versions}
\alias{update_all_versions}
\title{Helper functions for updating R/Py/Quarto/Posit versions}
\usage{
update_all_versions()
}
\description{
Currently this function would only be called by the package author, as it
updates inst/extdata/versions/*.csv
}
