% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_software_tibble.R
\name{update_software_csv}
\alias{update_software_csv}
\title{Helper function for updating R/Py/Quarto versions}
\usage{
update_software_csv()
}
\description{
Currently this function would only be called by the package author, as it
updates inst/extdata/versions/software.csv
}
