% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{create_config}
\alias{create_config}
\title{Create test config}
\usage{
create_config(file, pkg_name)
}
\arguments{
\item{file}{Config file name}

\item{pkg_name}{E.g. jrHealthCheckConnect description}
}
\description{
This function creates an example config. By default all tests are TRUE.
Some RSC servers intentionally won't have some features implemented.
}
\details{
If a test is missing from the config file, it is assume to be TRUE.
Therefore, the config file can be quite short and just list exceptions.
If the config file is missing, then all tests are carried out.

When merging configs, we either
* merge the new with existing
* force: overwrite existing file
* error: if a config file exists, raise an error
}
