% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posit_versions.R
\name{get_posit_versions}
\alias{get_posit_versions}
\title{Posit Versions}
\usage{
get_posit_versions(type = c("connect", "workbench", "drivers"), remote = TRUE)
}
\arguments{
\item{type}{Posit product of interest}

\item{remote}{Try scrapping NEWS first, but fall back to local if that fails}
}
\description{
Returns a tibble containing the Posit version, date of release,
and any associated CVEs
}
\examples{
get_posit_versions(type = "connect")

}
